Extract locally computed moon data into standalone moon_state_t
Moon rise/set/phase were stuffed into forecast.daily[0] by
enrichWithMoonData(), even though providers never supply moon data.
Compute moon_state_t once in setup() and pass it explicitly through the
render API instead. daily_t loses its moon fields.

+++ platformio/include/moon_tools.h
@@ -7,7 +7,7 @@
 typedef struct moon_state {
   time_t moonrise;
   time_t moonset;
-  double phase;  // 0.0 - 1.0
+  float phase;  // 0.0 - 1.0
 } moon_state_t;
 
 moon_state_t getMoonState(float latitude, float longitude);
